# refactor(batch): say goodbye to the state machine in NLJ