I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?
- Remove the state machine from NestedLoopJoinExecutor. Now each type of join is implemented by a nested loop, which is much more intuitive.
- ~~Add first-match optimization for left semi join.~~
- Add unit test for right outer join.
- Other changes to the utilities in NestedLoopJoinExecutor.

## Benchmark Reports (WIP)

*Left/RightAnti is ignored since it is almost the same as Left/RightSemi*.

| Type | Before | After |
| --- | --- | --- |
| Inner(32) | 361.26 ms | **332.64 ms** |
| Inner(128) | 297.71 ms | **270.09 ms** |
| Inner(512) | 280.10 ms | **264.01 ms** |
| LeftOuter(32) | 369.63 ms | **337.19 ms** |
| LeftOuter(128) | 300.46 ms | **270.30 ms** |
| LeftOuter(512) | 285.79 ms | **265.45 ms** |
| LeftSemi(32) | **5.0654 ms** | 277.09 ms (**4.8026 ms** *) |
| LeftSemi(128) | **14.785 ms** | 217.63 ms (**14.055 ms** *) |
| LeftSemi(512) | **53.397 ms** | 212.73 ms (**52.910 ms** *) |
| RightOuter(32) | 382.30 ms | **343.37 ms** |
| RightOuter(128) | 318.04 ms | **274.74 ms** |
| RightOuter(512) | 297.39 ms | **266.58 ms** |
| RightSemi(32) | 335.52 ms | **285.72 ms** |
| RightSemi(128) | 266.20 ms | **218.81 ms** |
| RightSemi(512) | 248.16 ms | **212.58 ms** |

*: with first-match optimization (WIP)

## Refer to a related PR or issue link (optional)
https://github.com/singularity-data/risingwave/issues/3662